requirements.txt
django~=5.0.4
djangorestframework~=3.15.1
psycopg2-binary

django-filter~=24.2